Scikit-Learn Update/CRISPR Usage Code
Fixed an issue where the Scikit-Learn didn't work with the lastest version of 0.20. The shift also needed the model to be retrained.

In addition, code for the CRIi/CRIa/CRIwt was placed into the program. However the value is currently hardcoded, needs to be able to adjusted via the GUI.