CRISPRi Screening
The tool is now able to be used for CRIi screening on Fosmids. This is done through finding ORFs and finding the guides within the ORFs and comparing those to Azimuth. In addition, the program also registers the nearest methanine.

Currently, the program seems to be running with no flaws. It is important to however test multiple edge cases to be sure. Also if some weirdo is reading these commits, let me know incase I can make my ORF finder faster. I feel like I should be able to.